backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.1.1",
    "fastapi>=0.135.1",
    "google-genai>=1.29.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]
